backend/requirements.txt
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
djoser==2.1.0
Pillow==9.2.0